First Draft Directory and Style completion
Completed First draft of Directory storage for each link.

All links seem to be working correctly except for beginner to master sections in each.

Styles applied to each section, although color palettes will need to be updated based on correct logo or main color to relate the feel to the corresponding section.